+++ src/components/body/SearchResult.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react';
 import ProviderContext from '../../context/ProviderContext';
 import GifList from './GifList';
-import { Gif } from '../../types/GifProvider';
+import { Gif } from '../../types/types';
 
 export interface SearchResultProps {
 	searchTerm: string;
