+++ src/components/body/TrendingResult.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useEffect, useState } from 'react';
 import ProviderContext from '../../context/ProviderContext';
 import GifList from './GifList';
-import { Gif } from '../../types/GifProvider';
+import { Gif } from '../../types/types';
 
 export interface TrendingResultProps {
 	columnsCount: number;
