+++ src/index.tsx
@@ -3,7 +3,9 @@ import ErrorBoundary from './components/ErrorBoundary';
 import GifPickerReact, { GifPickerReactProps } from './GifPickerReact';
 
 export { Theme } from './GifPickerReact';
-export type { Gif } from './types/GifProvider';
+export { GifProvider } from './types/GifProvider';
+export type { GifProviderAttribution, RegisterShareContext } from './types/GifProvider';
+export type { Gif, GifCategory, GifPreview } from './types/types';
 
 export interface GifPickerProps extends GifPickerReactProps {}
