+++ src/providers/tenor/TenorProvider.ts
@@ -1,11 +1,6 @@
 import { ContentFilter } from './tenor.types';
-import {
-	Gif,
-	GifProviderAttribution,
-	GifCategory,
-	GifProvider,
-	RegisterShareContext,
-} from '../../types/GifProvider';
+import { GifProvider, GifProviderAttribution, RegisterShareContext } from '../../types/GifProvider';
+import { Gif, GifCategory } from '../../types/types';
 import { TenorCategoriesResponse, TenorResult, TenorSearchResponse } from './tenor.types';
 
 const MEDIA_FILTER = 'gif,tinygif';
